backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
psycopg2-binary
python-multipart
python-jose
passlib[bcrypt]==1.7.4
shapely
bcrypt==3.2.0
PyYAML
python-dotenv
httpx


## requirements of the backend